requirements.txt
SQLAlchemy~=2.0.38
fastapi~=0.115.8
pydantic~=2.10.6
matplotlib~=3.8.4
pybit~=5.7.0
tabulate~=0.9.0
PyYAML~=6.0.2
uvicorn~=0.34.0
python-dotenv~=1.0.1
httpx~=0.28.1